// src/app/tabs/shared/models/season-peak-settings.model.ts
export class SeasonPeakSettings {
    id?: number;
    tariffId?: number;
    seasonId?: number;
    season?: string;
    weekType?: string;
    weekTypeId?: number;
    month?: string;
    monthId?: number;
    peakTypeId?: number;
    peakType?: string;
    dayId?: number;
    day?: string;
    hourStart?: string;
    hourEnd?: string;
    rate?: number;
    rowNumber?: number;
}